auto generate skill enum

+++ include/skill-system.h
@@ -43,7 +43,7 @@ void DisableUnitSkills(struct Unit * unit);
 /* Skill tetsers */
 bool SkillTesterBasic(struct Unit * unit, const u8 sid);
 bool SkillTesterFast(struct Unit * unit, const u8 sid);
-extern bool (* const SkillTester)(struct Unit * unit, const u8 sid);
+extern bool (* const SkillTester)(struct Unit * unit, const int sid);
 
 /* Prep equip skill list */
 struct PrepEquipSkillList {
